Renamed SwaggerRestServer to better named restClient
